requirements.txt
numpy==2.2.5
pandas==2.2.3
pandas_ta==0.3.14b0
pyotp==2.9.0
python-dotenv==1.1.0
SmartApi==1.1.0
smartapi_python==1.4.8
requests==2.31.0
Flask>=2.0
matplotlib==3.7.1
